+++ apps/web/src/components/RecipeDetailModal/RecipeHeroSection.tsx
@@ -30,6 +30,7 @@ interface RecipeHeroSectionProps {
   allTags: Tag[]
   onTagAdd: (tag: Tag) => void
   onTagRemove: (tagId: string) => void
+  onTagCreate: (name: string) => Promise<Tag>
   fileInputRef: RefObject<HTMLInputElement | null>
   onThumbnailFile: (e: ChangeEvent<HTMLInputElement>) => void
   imgUploading: boolean
@@ -50,6 +51,7 @@ const RecipeHeroSection = ({
   allTags,
   onTagAdd,
   onTagRemove,
+  onTagCreate,
   fileInputRef,
   onThumbnailFile,
   imgUploading,
@@ -75,6 +77,7 @@ const RecipeHeroSection = ({
         allTags={allTags}
         onAdd={onTagAdd}
         onRemove={onTagRemove}
+        onCreateTag={onTagCreate}
         editable={mode === 'editing'}
         addable
       />